renamed CLI.hpp to CLI.h; added CLI.cpp/.h to vcxproj file
